chapter_002/src/main/java/ru/job4j/tracker/ValidateInput.java
package ru.job4j.tracker;

public class ValidateInput implements Input {

    private final Input input;

    public ValidateInput(Input input) {
        this.input = input;
    }

    @Override
    public String askStr(String question) {
        return input.askStr(question);
    }

    @Override
    public int askInt(String question) {
        return Integer.valueOf(askStr(question));
    }

    @Override
    public int askInt(String question, int max) {
        boolean invalid = true;
        int value = -1;
        do {
            try {
                value = input.askInt(question, max);
                invalid = false;
            } catch (IllegalStateException e) {
                System.out.println("Please enter key from menu");
            } catch (NumberFormatException e) {
                System.out.println("Please enter validate data");
            }
        } while (invalid);
        return value;
    }
}
